API reference

String Scanners

SimpleStringScanners.SimpleStringScanner — Type

SimpleStringScanner

type for SimpleStringScanner

SimpleStringScanners.advance! — Function

advance!(scanner::SimpleStringScanner, width::Integer) -> SubString

Advances the inner pointer of scanner by width nodeunits, and returns scanned substring.

SimpleStringScanners.eos — Function

eos(scanner::SimpleStringScanner) -> Bool

Tests end-of-string.

parse

Base.parse — Function

parse(type::Type, s::SimpleStringScanner, width::Integer; base)

Variant of parse for SimpleStringScanner

SimpleStringScanners.parseBigFloat — Function

parseBigFloat(str::AbstractString; base)

Equivalent to

parse(BigFloat, str; base), and

parseBigFloat(str::AbstractString; base)

Equivalent to

parse(BigFloat, scanner, width; base).

SimpleStringScanners.parseBigInt — Function

parseBigInt(str::AbstractString; base)

Equivalent to

parse(BigInt, str; base), and

parseBigInt(str::AbstractString; base)

Equivalent to

parse(BigInt, scanner, width; base).

SimpleStringScanners.parseBool — Function

parseBool(str::AbstractString; base)

Equivalent to

parse(Bool, str; base), and

parseBool(str::AbstractString; base)

Equivalent to

parse(Bool, scanner, width; base).

SimpleStringScanners.parseFloat16 — Function

parseFloat16(str::AbstractString; base)

Equivalent to

parse(Float16, str; base), and

parseFloat16(str::AbstractString; base)

Equivalent to

parse(Float16, scanner, width; base).

SimpleStringScanners.parseFloat32 — Function

parseFloat32(str::AbstractString; base)

Equivalent to

parse(Float32, str; base), and

parseFloat32(str::AbstractString; base)

Equivalent to

parse(Float32, scanner, width; base).

SimpleStringScanners.parseFloat64 — Function

parseFloat64(str::AbstractString; base)

Equivalent to

parse(Float64, str; base), and

parseFloat64(str::AbstractString; base)

Equivalent to

parse(Float64, scanner, width; base).

SimpleStringScanners.parseInt8 — Function

parseInt8(str::AbstractString; base)

Equivalent to

parse(Int8, str; base), and

parseInt8(str::AbstractString; base)

Equivalent to

parse(Int8, scanner, width; base).

SimpleStringScanners.parseInt16 — Function

parseInt16(str::AbstractString; base)

Equivalent to

parse(Int16, str; base), and

parseInt16(str::AbstractString; base)

Equivalent to

parse(Int16, scanner, width; base).

SimpleStringScanners.parseInt32 — Function

parseInt32(str::AbstractString; base)

Equivalent to

parse(Int32, str; base), and

parseInt32(str::AbstractString; base)

Equivalent to

parse(Int32, scanner, width; base).

SimpleStringScanners.parseInt64 — Function

parseInt64(str::AbstractString; base)

Equivalent to

parse(Int64, str; base), and

parseInt64(str::AbstractString; base)

Equivalent to

parse(Int64, scanner, width; base).

SimpleStringScanners.parseInt128 — Function

parseInt128(str::AbstractString; base)

Equivalent to

parse(Int128, str; base), and

parseInt128(str::AbstractString; base)

Equivalent to

parse(Int128, scanner, width; base).

SimpleStringScanners.parseUInt8 — Function

parseUInt8(str::AbstractString; base)

Equivalent to

parse(UInt8, str; base), and

parseUInt8(str::AbstractString; base)

Equivalent to

parse(UInt8, scanner, width; base).

SimpleStringScanners.parseUInt16 — Function

parseUInt16(str::AbstractString; base)

Equivalent to

parse(UInt16, str; base), and

parseUInt16(str::AbstractString; base)

Equivalent to

parse(UInt16, scanner, width; base).

SimpleStringScanners.parseUInt32 — Function

parseUInt32(str::AbstractString; base)

Equivalent to

parse(UInt32, str; base), and

parseUInt32(str::AbstractString; base)

Equivalent to

parse(UInt32, scanner, width; base).

SimpleStringScanners.parseUInt64 — Function

parseUInt64(str::AbstractString; base)

Equivalent to

parse(UInt64, str; base), and

parseUInt64(str::AbstractString; base)

Equivalent to

parse(UInt64, scanner, width; base).

SimpleStringScanners.parseUInt128 — Function

parseUInt128(str::AbstractString; base)

Equivalent to

parse(UInt128, str; base), and

parseUInt128(str::AbstractString; base)

Equivalent to

parse(UInt128, scanner, width; base).

tryparse

Base.tryparse — Function

tryparse(type::Type, s::SimpleStringScanner, width::Integer; base)

Variant of tryparse for SimpleStringScanner

SimpleStringScanners.tryparseBigFloat — Function

tryparseBigFloat(str::AbstractString; base)

Equivalent to

tryparse(BigFloat, str; base), and

tryparseBigFloat(str::AbstractString; base)

Equivalent to

tryparse(BigFloat, scanner, width; base).

SimpleStringScanners.tryparseBigInt — Function

tryparseBigInt(str::AbstractString; base)

Equivalent to

tryparse(BigInt, str; base), and

tryparseBigInt(str::AbstractString; base)

Equivalent to

tryparse(BigInt, scanner, width; base).

SimpleStringScanners.tryparseBool — Function

tryparseBool(str::AbstractString; base)

Equivalent to

tryparse(Bool, str; base), and

tryparseBool(str::AbstractString; base)

Equivalent to

tryparse(Bool, scanner, width; base).

SimpleStringScanners.tryparseFloat16 — Function

tryparseFloat16(str::AbstractString; base)

Equivalent to

tryparse(Float16, str; base), and

tryparseFloat16(str::AbstractString; base)

Equivalent to

tryparse(Float16, scanner, width; base).

SimpleStringScanners.tryparseFloat32 — Function

tryparseFloat32(str::AbstractString; base)

Equivalent to

tryparse(Float32, str; base), and

tryparseFloat32(str::AbstractString; base)

Equivalent to

tryparse(Float32, scanner, width; base).

SimpleStringScanners.tryparseFloat64 — Function

tryparseFloat64(str::AbstractString; base)

Equivalent to

tryparse(Float64, str; base), and

tryparseFloat64(str::AbstractString; base)

Equivalent to

tryparse(Float64, scanner, width; base).

SimpleStringScanners.tryparseInt8 — Function

tryparseInt8(str::AbstractString; base)

Equivalent to

tryparse(Int8, str; base), and

tryparseInt8(str::AbstractString; base)

Equivalent to

tryparse(Int8, scanner, width; base).

SimpleStringScanners.tryparseInt16 — Function

tryparseInt16(str::AbstractString; base)

Equivalent to

tryparse(Int16, str; base), and

tryparseInt16(str::AbstractString; base)

Equivalent to

tryparse(Int16, scanner, width; base).

SimpleStringScanners.tryparseInt32 — Function

tryparseInt32(str::AbstractString; base)

Equivalent to

tryparse(Int32, str; base), and

tryparseInt32(str::AbstractString; base)

Equivalent to

tryparse(Int32, scanner, width; base).

SimpleStringScanners.tryparseInt64 — Function

tryparseInt64(str::AbstractString; base)

Equivalent to

tryparse(Int64, str; base), and

tryparseInt64(str::AbstractString; base)

Equivalent to

tryparse(Int64, scanner, width; base).

SimpleStringScanners.tryparseInt128 — Function

tryparseInt128(str::AbstractString; base)

Equivalent to

tryparse(Int128, str; base), and

tryparseInt128(str::AbstractString; base)

Equivalent to

tryparse(Int128, scanner, width; base).

SimpleStringScanners.tryparseUInt8 — Function

tryparseUInt8(str::AbstractString; base)

Equivalent to

tryparse(UInt8, str; base), and

tryparseUInt8(str::AbstractString; base)

Equivalent to

tryparse(UInt8, scanner, width; base).

SimpleStringScanners.tryparseUInt16 — Function

tryparseUInt16(str::AbstractString; base)

Equivalent to

tryparse(UInt16, str; base), and

tryparseUInt16(str::AbstractString; base)

Equivalent to

tryparse(UInt16, scanner, width; base).

SimpleStringScanners.tryparseUInt32 — Function

tryparseUInt32(str::AbstractString; base)

Equivalent to

tryparse(UInt32, str; base), and

tryparseUInt32(str::AbstractString; base)

Equivalent to

tryparse(UInt32, scanner, width; base).

SimpleStringScanners.tryparseUInt64 — Function

tryparseUInt64(str::AbstractString; base)

Equivalent to

tryparse(UInt64, str; base), and

tryparseUInt64(str::AbstractString; base)

Equivalent to

tryparse(UInt64, scanner, width; base).

SimpleStringScanners.tryparseUInt128 — Function

tryparseUInt128(str::AbstractString; base)

Equivalent to

tryparse(UInt128, str; base), and

tryparseUInt128(str::AbstractString; base)

Equivalent to

tryparse(UInt128, scanner, width; base).

parse1

SimpleStringScanners.parse1 — Function

parse1(type::Type, str::AbstractString; base)

Variant of Base.parse that can handle Fortran's floating point numbers whose exponent part starts with d character. Substitutes d and D in str by e's, and then invokes Base.parse.

parse1(type::Type, s::SimpleStringScanner, width::Integer; base)

Variant of parse1 for SimpleStringScanner

SimpleStringScanners.parse1BigFloat — Function

parse1BigFloat(str::AbstractString; base)

Equivalent to

parse1(BigFloat, str; base), and

parse1BigFloat(str::AbstractString; base)

Equivalent to

parse1(BigFloat, scanner, width; base).

SimpleStringScanners.parse1BigInt — Function

parse1BigInt(str::AbstractString; base)

Equivalent to

parse1(BigInt, str; base), and

parse1BigInt(str::AbstractString; base)

Equivalent to

parse1(BigInt, scanner, width; base).

SimpleStringScanners.parse1Bool — Function

parse1Bool(str::AbstractString; base)

Equivalent to

parse1(Bool, str; base), and

parse1Bool(str::AbstractString; base)

Equivalent to

parse1(Bool, scanner, width; base).

SimpleStringScanners.parse1Float16 — Function

parse1Float16(str::AbstractString; base)

Equivalent to

parse1(Float16, str; base), and

parse1Float16(str::AbstractString; base)

Equivalent to

parse1(Float16, scanner, width; base).

SimpleStringScanners.parse1Float32 — Function

parse1Float32(str::AbstractString; base)

Equivalent to

parse1(Float32, str; base), and

parse1Float32(str::AbstractString; base)

Equivalent to

parse1(Float32, scanner, width; base).

SimpleStringScanners.parse1Float64 — Function

parse1Float64(str::AbstractString; base)

Equivalent to

parse1(Float64, str; base), and

parse1Float64(str::AbstractString; base)

Equivalent to

parse1(Float64, scanner, width; base).

SimpleStringScanners.parse1Int8 — Function

parse1Int8(str::AbstractString; base)

Equivalent to

parse1(Int8, str; base), and

parse1Int8(str::AbstractString; base)

Equivalent to

parse1(Int8, scanner, width; base).

SimpleStringScanners.parse1Int16 — Function

parse1Int16(str::AbstractString; base)

Equivalent to

parse1(Int16, str; base), and

parse1Int16(str::AbstractString; base)

Equivalent to

parse1(Int16, scanner, width; base).

SimpleStringScanners.parse1Int32 — Function

parse1Int32(str::AbstractString; base)

Equivalent to

parse1(Int32, str; base), and

parse1Int32(str::AbstractString; base)

Equivalent to

parse1(Int32, scanner, width; base).

SimpleStringScanners.parse1Int64 — Function

parse1Int64(str::AbstractString; base)

Equivalent to

parse1(Int64, str; base), and

parse1Int64(str::AbstractString; base)

Equivalent to

parse1(Int64, scanner, width; base).

SimpleStringScanners.parse1Int128 — Function

parse1Int128(str::AbstractString; base)

Equivalent to

parse1(Int128, str; base), and

parse1Int128(str::AbstractString; base)

Equivalent to

parse1(Int128, scanner, width; base).

SimpleStringScanners.parse1UInt8 — Function

parse1UInt8(str::AbstractString; base)

Equivalent to

parse1(UInt8, str; base), and

parse1UInt8(str::AbstractString; base)

Equivalent to

parse1(UInt8, scanner, width; base).

SimpleStringScanners.parse1UInt16 — Function

parse1UInt16(str::AbstractString; base)

Equivalent to

parse1(UInt16, str; base), and

parse1UInt16(str::AbstractString; base)

Equivalent to

parse1(UInt16, scanner, width; base).

SimpleStringScanners.parse1UInt32 — Function

parse1UInt32(str::AbstractString; base)

Equivalent to

parse1(UInt32, str; base), and

parse1UInt32(str::AbstractString; base)

Equivalent to

parse1(UInt32, scanner, width; base).

SimpleStringScanners.parse1UInt64 — Function

parse1UInt64(str::AbstractString; base)

Equivalent to

parse1(UInt64, str; base), and

parse1UInt64(str::AbstractString; base)

Equivalent to

parse1(UInt64, scanner, width; base).

SimpleStringScanners.parse1UInt128 — Function

parse1UInt128(str::AbstractString; base)

Equivalent to

parse1(UInt128, str; base), and

parse1UInt128(str::AbstractString; base)

Equivalent to

parse1(UInt128, scanner, width; base).

tryparse1

SimpleStringScanners.tryparse1 — Function

tryparse1(type::Type, str::AbstractString; base)

Variant of Base.tryparse that can handle Fortran's floating point numbers whose exponent part starts with d character. Substitutes d and D in str by e's, and then invokes Base.tryparse.

tryparse1(type::Type, s::SimpleStringScanner, width::Integer; base)

Variant of tryparse1 for SimpleStringScanner

SimpleStringScanners.tryparse1BigFloat — Function

tryparse1BigFloat(str::AbstractString; base)

Equivalent to

tryparse1(BigFloat, str; base), and

tryparse1BigFloat(str::AbstractString; base)

Equivalent to

tryparse1(BigFloat, scanner, width; base).

SimpleStringScanners.tryparse1BigInt — Function

tryparse1BigInt(str::AbstractString; base)

Equivalent to

tryparse1(BigInt, str; base), and

tryparse1BigInt(str::AbstractString; base)

Equivalent to

tryparse1(BigInt, scanner, width; base).

SimpleStringScanners.tryparse1Bool — Function

tryparse1Bool(str::AbstractString; base)

Equivalent to

tryparse1(Bool, str; base), and

tryparse1Bool(str::AbstractString; base)

Equivalent to

tryparse1(Bool, scanner, width; base).

SimpleStringScanners.tryparse1Float16 — Function

tryparse1Float16(str::AbstractString; base)

Equivalent to

tryparse1(Float16, str; base), and

tryparse1Float16(str::AbstractString; base)

Equivalent to

tryparse1(Float16, scanner, width; base).

SimpleStringScanners.tryparse1Float32 — Function

tryparse1Float32(str::AbstractString; base)

Equivalent to

tryparse1(Float32, str; base), and

tryparse1Float32(str::AbstractString; base)

Equivalent to

tryparse1(Float32, scanner, width; base).

SimpleStringScanners.tryparse1Float64 — Function

tryparse1Float64(str::AbstractString; base)

Equivalent to

tryparse1(Float64, str; base), and

tryparse1Float64(str::AbstractString; base)

Equivalent to

tryparse1(Float64, scanner, width; base).

SimpleStringScanners.tryparse1Int8 — Function

tryparse1Int8(str::AbstractString; base)

Equivalent to

tryparse1(Int8, str; base), and

tryparse1Int8(str::AbstractString; base)

Equivalent to

tryparse1(Int8, scanner, width; base).

SimpleStringScanners.tryparse1Int16 — Function

tryparse1Int16(str::AbstractString; base)

Equivalent to

tryparse1(Int16, str; base), and

tryparse1Int16(str::AbstractString; base)

Equivalent to

tryparse1(Int16, scanner, width; base).

SimpleStringScanners.tryparse1Int32 — Function

tryparse1Int32(str::AbstractString; base)

Equivalent to

tryparse1(Int32, str; base), and

tryparse1Int32(str::AbstractString; base)

Equivalent to

tryparse1(Int32, scanner, width; base).

SimpleStringScanners.tryparse1Int64 — Function

tryparse1Int64(str::AbstractString; base)

Equivalent to

tryparse1(Int64, str; base), and

tryparse1Int64(str::AbstractString; base)

Equivalent to

tryparse1(Int64, scanner, width; base).

SimpleStringScanners.tryparse1Int128 — Function

tryparse1Int128(str::AbstractString; base)

Equivalent to

tryparse1(Int128, str; base), and

tryparse1Int128(str::AbstractString; base)

Equivalent to

tryparse1(Int128, scanner, width; base).

SimpleStringScanners.tryparse1UInt8 — Function

tryparse1UInt8(str::AbstractString; base)

Equivalent to

tryparse1(UInt8, str; base), and

tryparse1UInt8(str::AbstractString; base)

Equivalent to

tryparse1(UInt8, scanner, width; base).

SimpleStringScanners.tryparse1UInt16 — Function

tryparse1UInt16(str::AbstractString; base)

Equivalent to

tryparse1(UInt16, str; base), and

tryparse1UInt16(str::AbstractString; base)

Equivalent to

tryparse1(UInt16, scanner, width; base).

SimpleStringScanners.tryparse1UInt32 — Function

tryparse1UInt32(str::AbstractString; base)

Equivalent to

tryparse1(UInt32, str; base), and

tryparse1UInt32(str::AbstractString; base)

Equivalent to

tryparse1(UInt32, scanner, width; base).

SimpleStringScanners.tryparse1UInt64 — Function

tryparse1UInt64(str::AbstractString; base)

Equivalent to

tryparse1(UInt64, str; base), and

tryparse1UInt64(str::AbstractString; base)

Equivalent to

tryparse1(UInt64, scanner, width; base).

SimpleStringScanners.tryparse1UInt128 — Function

tryparse1UInt128(str::AbstractString; base)

Equivalent to

tryparse1(UInt128, str; base), and

tryparse1UInt128(str::AbstractString; base)

Equivalent to

tryparse1(UInt128, scanner, width; base).